+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## 0.16.0
+
+### Minor Changes
+
+- Add dndRootElement for drag and drop functionality
+
 ## 0.15.0
 
 ### Minor Changes
@@ -242,7 +248,8 @@
 - Add custom cunningham.ts file
 - Still a WIP version
 
-[unreleased]: https://github.com/suitenumerique/ui-kit/compare/v0.15.0...main
+[unreleased]: https://github.com/suitenumerique/ui-kit/compare/v0.16.0...main
+[0.16.0]: https://github.com/suitenumerique/ui-kit/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/suitenumerique/ui-kit/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/suitenumerique/ui-kit/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/suitenumerique/ui-kit/compare/v0.12.0...v0.13.0